moved the location of the 'serverAddress' file
